requirements.txt
azure-storage-blob==12.3.0
azure-batch==9.0.0

